backend/multimodal_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pillow
ffmpeg-python
openai
